+++ cmd/util.go
@@ -72,7 +72,7 @@ func filter(options []string, tag string) (commands []string, err error) {
 		snippetTexts[t] = s
 		if config.Flag.Color {
 			t = fmt.Sprintf("[%s]: %s%s",
-				color.RedString(s.Description), command, color.BlueString(tags))
+				color.GreenString(s.Description), command, color.BlueString(tags))
 		}
 		text += t + "\n"
 	}