+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # QOwnNotes command-line snippet manager changelog
 
 ## v0.6.0
+- Add support for storing commands in [Atuin](https://atuin.sh/) on execution
+  when using the `--atuin` flag (for [#15](https://github.com/qownnotes/qc/issues/15))
 - Update dependencies
 
 ## v0.5.1